Lazy-load menu item controls
Don't render menu item controls until the containing menu section is
opened for the first time. While this requires a minimal custom section
for some of the JS and a couple of initially odd changes to the custom
control JS, this implementation should work well and greatly improve
performance. With this, menus won't really have much of an impact on
Customizer load time at all.

Fixes #22.