Debounce available-item search calls
This limits the ajax calls to only occur when the user stops typing for
at least 250 ms. Also avoids an extra ajax call to looks for more
results in many cases where there's less than one page of results. Clean
up cases where the search term is changed before results come back, as
well as cleaning up usage of loading indicators based on when results
are actually loading or not.