Add menu location checkboxes to each menu section
The values are cross-synced with the menu location controls in the menu
location section and the UI is updated for both accordingly. Note that
the checkboxes are not actual controls, as the associated settings are
handled by the menu location (select) controls. The currently-set menu
name is displayed next to the checkboxes if the current menu is not set
to that location.