Lazy-load available menu items a bit earlier
Load the first batch of available menu items for each type when the
menus panel is first expanded, so that the items are generally done
loading by the time the user gets to the add-item panel, even if they go
straight there. Will only slow down initial load if menus or a child
section/control is autofocused, but we'd want menu items loaded in that
case anyway.