Restore lazy loading of menu item controls
This is critical for ensuring that menus don't slow down the Customizer,
especially on initial load. Menu item controls are rendered to html and
inserted into the DOM as-needed, one menu at a time when the menu
section is first opened (a nearly-instant process when handled one menu
at a time like this).